+++ main.js
@@ -2,10 +2,12 @@ var express = require('express');
 var app = express();
 var authRouter = require('./routes/auth');
 var categoryRouter = require('./routes/category');
+var cardRouter = require('./routes/card');
 
 app.use(express.json());
 app.use('/auth', authRouter);
 app.use('/category', categoryRouter);
+app.use('/category/card', cardRouter);
 
 app.use(function(req, res, next) {
     res.status(404).send('Sorry cant find that!');
